pyproject.toml
[project]
name = "aarlp"
version = "1.0.0"
description = "AI-Agentic Recruitment Lifecycle Platform"
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}
authors = [
    {name = "AARLP Team"}
]
keywords = ["recruitment", "ai", "langgraph", "crewai", "fastapi"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Framework :: FastAPI",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "langgraph>=0.0.40",
    "langchain>=0.1.0",
    "langchain-openai>=0.0.5",
    "langgraph-checkpoint-postgres>=0.0.1",
    "crewai>=0.1.0",
    "asyncpg>=0.29.0",
    "sqlalchemy[asyncio]>=2.0.0",
    "pgvector>=0.2.0",
    "openai>=1.10.0",
    "twilio>=8.10.0",
    "elevenlabs>=0.2.0",
    "google-api-python-client>=2.100.0",
    "google-auth-httplib2>=0.1.1",
    "google-auth-oauthlib>=1.1.0",
    "pypdf2>=3.0.0",
    "pdfplumber>=0.10.0",
    "sentence-transformers>=2.2.0",
    "python-dotenv>=1.0.0",
    "httpx>=0.26.0",
    "numpy>=1.26.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.1.0",
    "mypy>=1.8.0",
    "ruff>=0.1.0",
    "black>=24.1.0",
    "isort>=5.13.0",
    "pre-commit>=3.6.0",
]

[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["app*", "tests*"]

# ============================================================================
# Ruff Configuration (Fast Python Linter)
# ============================================================================
[tool.ruff]
target-version = "py311"
line-length = 88
select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # Pyflakes
    "I",      # isort
    "B",      # flake8-bugbear
    "C4",     # flake8-comprehensions
    "UP",     # pyupgrade
    "ARG",    # flake8-unused-arguments
    "SIM",    # flake8-simplify
    "TCH",    # flake8-type-checking
    "PTH",    # flake8-use-pathlib
    "ERA",    # eradicate (commented-out code)
    "RUF",    # Ruff-specific rules
]
ignore = [
    "E501",   # line too long (handled by black)
    "B008",   # do not perform function calls in argument defaults
    "B904",   # raise without from (we handle this case-by-case)
    "ARG001", # unused function arguments (needed for API signatures)
]

[tool.ruff.per-file-ignores]
"tests/*" = ["ARG", "S101"]  # Allow assert and unused args in tests
"__init__.py" = ["F401"]      # Allow unused imports in __init__

[tool.ruff.isort]
known-first-party = ["app"]
force-single-line = false
lines-after-imports = 2

# ============================================================================
# Black Configuration (Code Formatter)
# ============================================================================
[tool.black]
line-length = 88
target-version = ["py311"]
include = '\.pyi?$'
exclude = '''
/(
    \.git
    | \.mypy_cache
    | \.ruff_cache
    | \.venv
    | venv
    | __pycache__
    | build
    | dist
)/
'''

# ============================================================================
# isort Configuration
# ============================================================================
[tool.isort]
profile = "black"
line_length = 88
known_first_party = ["app"]
skip = [".venv", "venv"]

# ============================================================================
# MyPy Configuration (Type Checking)
# ============================================================================
[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
warn_unused_ignores = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_configs = true

# Per-module configuration
[[tool.mypy.overrides]]
module = [
    "twilio.*",
    "elevenlabs.*",
    "pdfplumber.*",
    "crewai.*",
    "langgraph.*",
    "langchain.*",
    "pgvector.*",
    "google.*",
]
ignore_missing_imports = true

# ============================================================================
# Pytest Configuration
# ============================================================================
[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
python_functions = ["test_*"]
asyncio_mode = "auto"
addopts = [
    "-v",
    "--tb=short",
    "--strict-markers",
    "-ra",
]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests as integration tests",
]
filterwarnings = [
    "ignore::DeprecationWarning",
]

# ============================================================================
# Coverage Configuration
# ============================================================================
[tool.coverage.run]
source = ["app"]
branch = true
omit = [
    "*/tests/*",
    "*/__init__.py",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise NotImplementedError",
    "if TYPE_CHECKING:",
    "if __name__ == .__main__.:",
]
fail_under = 70
show_missing = true
